+++ src/main/java/konkuk/corkCharge/global/api/GJRestaurantOpenApi/dto/RestaurantOpenApiResponse.java
@@ -1,11 +1,9 @@
-package konkuk.corkCharge.global.openApi.GJRestaurantOpenApi.dto;
+package konkuk.corkCharge.global.api.GJRestaurantOpenApi.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @Getter
 @NoArgsConstructor
